requirements.txt
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.2
binaryornot==0.4.4
Brotli==1.1.0
celery==5.5.3
certifi==2024.8.30
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.6.0
cssselect2==0.8.0
diff-match-patch==20230430
distro==1.9.0
Django==5.1.1
django-browser-reload==1.15.0
django-cors-headers==4.6.0
django-cotton==1.0.9
django-import-export==4.1.1
django-tailwind==3.8.0
django-unfold==0.38.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
fonttools==4.60.1
idna==3.8
Jinja2==3.1.4
kombu==5.5.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysqlclient==2.2.7
numpy==2.1.1
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==12.0.0
prompt_toolkit==3.0.52
pycparser==2.23
pydyf==0.11.0
Pygments==2.18.0
pyotp==2.9.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.3
rich==13.8.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tablib==3.5.0
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
types-python-dateutil==2.9.0.20240821
tzdata==2025.2
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
openpyxl>=3.0.10
